moved the CharacterBase class under the character module and merged it with the Character class
