+++ physics_platformer/src/physics_platformer/animation/animation_actor.py
@@ -202,7 +202,7 @@ class AnimationActor(SpriteAnimator):
     self.rigid_body_np_ = NodePath(BulletRigidBodyNode('RigidBody'))
     rigid_body = self.rigid_body_np_.node()
     rigid_body.setIntoCollideMask(CollisionMasks.RIGID_BODY)
-    rigid_body.setDeactivationEnabled(False,True)
+    rigid_body.setDeactivationEnabled(False)
     
     # collection all boxes
     collision_boxes = []
