+++ physics_platformer/src/physics_platformer/character/__init__.py
@@ -1,3 +1,3 @@
-from character_state_machine import *
+from character_info import *
 from character_states import *
-from character import *
+from character_base import *
