+++ physics_platformer/src/physics_platformer/character/character_base.py
@@ -1,18 +1,27 @@
-
-from panda3d.core import Vec3
-from physics_platformer.collision import CollisionMasks
-from physics_platformer.sprite import SpriteAnimator
-from physics_platformer.animation import AnimationActor
-from physics_platformer.game_object import CharacterInfo
-from physics_platformer.game_object import AnimationSpriteAlignment
-from physics_platformer.game_object import AnimatableObject
+import logging
 from panda3d.core import NodePath
 from panda3d.core import TransformState
 from panda3d.bullet import BulletBoxShape
 from panda3d.bullet import BulletGenericConstraint
 from panda3d.bullet import BulletWorld
 from docutils import TransformSpec
-import logging
+from panda3d.core import Vec3
+
+from physics_platformer.collision import CollisionMasks
+from physics_platformer.sprite import SpriteAnimator
+from physics_platformer.animation import AnimationActor
+from physics_platformer.character import CharacterInfo
+from physics_platformer.game_object import AnimationSpriteAlignment
+from physics_platformer.game_object import AnimatableObject
+from physics_platformer.character.character_states import *
+from physics_platformer.state_machine import Action
+from physics_platformer.state_machine import State
+from physics_platformer.state_machine import StateMachine
+from physics_platformer.state_machine import StateEvent
+from physics_platformer.state_machine import StateMachineActions
+from physics_platformer.game_actions import *
+from physics_platformer.character.character_states import CharacterStateKeys
+from physics_platformer.character.character_states import CharacterStates
 
 
 class CharacterBase(AnimatableObject):
@@ -21,17 +30,18 @@ class CharacterBase(AnimatableObject):
   
   def  __init__(self,info):
     
-    self.character_info_ = info
+    # animation base setup
+    self.character_info_ = info   
+    size = Vec3(info.width, AnimationActor.DEFAULT_WIDTH , info.height)
+    AnimatableObject.__init__(self,info.name,size,info.mass)    
+    self.animation_root_np_.setPos(Vec3(0,0,0))
     
-    # turning active rigid bodies left and right
+        # constraints
     self.left_constraint_ = None
     self.right_constraint_ = None
     self.active_constraint_ = None
     
-    size = Vec3(info.width, AnimationActor.DEFAULT_WIDTH , info.height)
-    AnimatableObject.__init__(self,info.name,size,info.mass)
-    
-    # re-creating shape
+    # rigid body
     shapes = self.node().getShapes()
     for shape in shapes:
       self.node().removeShape(shape)
@@ -41,11 +51,88 @@ class CharacterBase(AnimatableObject):
     self.node().setMass(self.character_info_.mass)
     self.node().setLinearFactor((1,0,1))   
     self.node().setAngularFactor((0,0,0)) 
-    self.node().setDeactivationEnabled(False,True)
-    self.setCollideMask(CollisionMasks.NO_COLLISION)
+    self.node().setDeactivationEnabled(False)
+    self.setCollideMask(CollisionMasks.NO_COLLISION)    
     
-    self.animation_root_np_.setPos(Vec3(0,0,0))
+    # state machine
+    self.sm_ = StateMachine()   
+    
+  def setup(self):    
 
+    self.__setupDefaultStates__()
+    self.__setupTransitionRules__()
+    
+    return True
+  
+  def clampBottom(self,z):
+    '''
+    Character.clampDown(double )
+      Sets the bottom z value of the character
+    '''
+    
+    # setting vertical speed to zero
+    vel = self.node().getLinearVelocity()
+    vel.setZ(0)
+    self.node().setLinearVelocity(vel)
+    
+    # clamping to platform surface    
+    bbox = self.getAnimatorActor().getRigidBodyBoundingBox()  
+    self.setZ(z - bbox.bottom) 
+  
+  def clampTop(self,z):
+    '''
+    Character.clampDown(double )
+      Sets the Top z value of the character
+    '''
+    
+    # setting vertical speed to zero
+    vel = self.node().getLinearVelocity()
+    vel.setZ(0)
+    self.node().setLinearVelocity(vel)
+    
+    # clamping to platform surface    
+    bbox = self.getAnimatorActor().getRigidBodyBoundingBox()  
+    self.setZ(z - bbox.top) 
+    
+  def execute(self,action):
+    self.sm_.execute(action)
+    
+  def update(self,dt):
+    #self.node().setActive(True)
+    self.sm_.execute(GeneralActions.GameStep(dt))
+    
+  def __setupDefaultStates__(self):
+    
+    # creating default states
+    standing_state = CharacterStates.StandingState(self, self.sm_)
+    running_state = CharacterStates.RunningState(self,self.sm_)
+    takeoff_state = CharacterStates.TakeoffState(self,self.sm_)
+    jump_state = CharacterStates.JumpState(self,self.sm_)
+    fall_state = CharacterStates.FallState(self,self.sm_)
+    land_state = CharacterStates.LandState(self,self.sm_)
+        
+    self.sm_.addState(fall_state)
+    self.sm_.addState(standing_state)
+    self.sm_.addState(running_state)
+    self.sm_.addState(takeoff_state)
+    self.sm_.addState(jump_state)
+    self.sm_.addState(land_state)    
+    
+  def __setupTransitionRules__(self):    
+    
+    self.sm_.addTransition(CharacterStateKeys.FALLING, CollisionAction.SURFACE_COLLISION, CharacterStateKeys.LANDING)
+    self.sm_.addTransition(CharacterStateKeys.STANDING,CharacterActions.MOVE_RIGHT.key,CharacterStateKeys.RUNNING)
+    self.sm_.addTransition(CharacterStateKeys.STANDING,CharacterActions.MOVE_LEFT.key,CharacterStateKeys.RUNNING)
+    self.sm_.addTransition(CharacterStateKeys.STANDING,CharacterActions.JUMP.key,CharacterStateKeys.TAKEOFF)
+    self.sm_.addTransition(CharacterStateKeys.STANDING,CharacterActions.FALL.key,CharacterStateKeys.FALLING)
+    
+    self.sm_.addTransition(CharacterStateKeys.RUNNING,CharacterActions.FALL.key,CharacterStateKeys.FALLING)
+    self.sm_.addTransition(CharacterStateKeys.RUNNING,CharacterActions.JUMP.key,CharacterStateKeys.TAKEOFF)
+    self.sm_.addTransition(CharacterStateKeys.RUNNING,CharacterActions.MOVE_NONE.key,CharacterStateKeys.STANDING)
+    
+    self.sm_.addTransition(CharacterStateKeys.TAKEOFF, StateMachineActions.DONE.key, CharacterStateKeys.JUMPING)
+    self.sm_.addTransition(CharacterStateKeys.JUMPING, StateMachineActions.DONE.key, CharacterStateKeys.FALLING)
+    self.sm_.addTransition(CharacterStateKeys.LANDING, StateMachineActions.DONE.key, CharacterStateKeys.STANDING)
     
   def addAnimationActor(self,name,anim_actor,
                         align = (AnimationSpriteAlignment.CENTER_OFFSET_ALIGN),
