+++ physics_platformer/src/physics_platformer/character/character_info.py
