+++ physics_platformer/src/physics_platformer/game_object/__init__.py
@@ -1,4 +1,2 @@
-from character_info import *
 from game_object import *
 from animatable_object import *
-from character_base import *
